Refactor javascript usage.
