+++ app/javascript/controllers/posts.js
@@ -1,0 +1,7 @@
+
+
+export default {
+    index: _ => {
+        console.log("imported index!")
+    }
+}
